Folder strcuture changes
Added the deleted files

Data mapping from machine to device added few properties as per the discussion

Changes done as per PR comments
